mega update

+++ CreoHub.Application/Commands/StorageCommands/DeleteFile.cs
@@ -26,14 +26,14 @@ public class DeleteFileHandler : IRequestHandler<DeleteFile, BaseResponse<bool>>
         try
         {
             var storageObject = await _storageObjectRepository.GetByIdAsync(request.id);
-            if (storageObject.FileType != FileType.Unregistred)
-            {
-                return BaseResponse<bool>.Fail("Файл используется в качестве контента или медиа. Отвяжите перед удалением");
-            }
+            if (storageObject == null)
+                return BaseResponse<bool>.Fail("Файл не найден");
             if (storageObject.OwnerId != request.shopId)
-            {
                 return BaseResponse<bool>.Fail("Вы не являетесь владельцем данного файла");
-            }
+            if (storageObject.IsSystemLocked)
+                return BaseResponse<bool>.Fail("Файл нельзя удалить — он содержит данные покупателей");
+            if (storageObject.FileType != FileType.Unregistred)
+                return BaseResponse<bool>.Fail("Файл используется в качестве контента или медиа. Отвяжите перед удалением");
             _storageObjectRepository.Remove(storageObject);
             await _unitOfWork.SaveChangesAsync(cancellationToken);
             await _storageService.DeleteFileAsync(storageObject.Key);
